Get the interface of the user works
The user interface works to query and select information from the database. Just was modified the line 3 using GET to pass the data entered by the user to the query line, and modify the line 384 to use GET as method instead POST. To pass data entered by user only use thre first space named Firstname writing the bame of the restaurant, tested with TOFU TOFU, CAFE MANHATTAN, and PIO PIO, worked with each of these restaurants.