Downloaded Database in CSV format
This file is the CSV download format of the database for NY restaurant inspections in Manhattan. First activate xampp-control-panel , start Apache and MySQL modules, type in your browser localhost/phpmyadmin , create the database named: nyc_restaurant_inspection , then create a table with name: nyc_restaurants , the format of the table and name of the columns you can see it opening the database file using Visual Studio, once created the table just click on icon IMPORT and select this CSV file, if stoped during the import just try it again. Once finished the import we are ready to practice with the functionality of the web page.